Split email update into separate providers controller
