+++ app/controllers/providers/applicants_controller.rb
@@ -18,24 +18,6 @@ module Providers
       end
     end
 
-    def edit
-      @form = Applicants::EmailForm.new(model: applicant)
-    end
-
-    def update
-      @form = Applicants::EmailForm.new(edit_params)
-
-      if @form.save
-        redirect_to(
-          action_for_next_step(options: { application: legal_aid_application, applicant: @form.model }),
-          # TODO: Remove this - currently just a way of displaying a usable link
-          notice: "Email link will be to: #{citizens_legal_aid_application_url(legal_aid_application.generate_secure_id)}"
-        )
-      else
-        render :edit
-      end
-    end
-
     private
 
     def applicant_params
@@ -46,20 +28,8 @@ module Providers
       applicant_params.merge(legal_aid_application_id: params[:legal_aid_application_id])
     end
 
-    def edit_params
-      email_params.merge(model: applicant)
-    end
-
-    def email_params
-      params.require(:applicant).permit(:email)
-    end
-
-    def applicant
-      @applicant ||= legal_aid_application.applicant
-    end
-
     def set_current_step
-      @current_step = %w[edit update].include?(params[:action]) ? :email : :basic_details
+      @current_step = :basic_details
     end
   end
 end
