+++ app/views/providers/emails/show.html.erb
@@ -12,7 +12,7 @@
       </h1>
     </legend>
 
-    <%= form_for(@form, url: providers_legal_aid_application_applicant_path(@legal_aid_application), method: 'put') do |form| %>
+    <%= form_for(@form, url: providers_legal_aid_application_email_path(@legal_aid_application), method: 'put') do |form| %>
 
       <div class="govuk-!-padding-bottom-6"></div>
