+++ config/routes.rb
@@ -27,11 +27,12 @@ Rails.application.routes.draw do
     root to: 'start#index'
 
     resources :legal_aid_applications, path: 'applications', only: %i[index new create] do
-      resource :applicant
+      resource :applicant, only: %i[new create]
 
       resources :check_your_answers, only: [:index] do
         post :confirm, on: :collection
       end
+      resource :email, only: %i[show update]
 
       resources :check_benefits, only: [:index]
     end
